backend/backendservice/requirements.txt
asgiref==3.8.1
certifi==2024.7.4
charset-normalizer==3.3.2
Django==4.1.5
django-cors-headers==3.13.0
django-mysql==4.8.0
django-stubs==5.0.4
django-stubs-ext==5.0.4
djangorestframework==3.12.0
djangorestframework-stubs==3.15.0
idna==3.7
mysqlclient==2.1.1
pika==1.3.1
pytz==2022.7.1
requests==2.32.3
setuptools==72.1.0
sqlparse==0.5.1
types-PyYAML==6.0.12.20240724
types-requests==2.32.0.20240712
typing_extensions==4.12.2
urllib3==2.2.2
wheel==0.43.0
